requirements.txt
Flask
cohere
python-dotenv
requests